refactor: opt out from suspense and use loading convention from nexjs instead
